Update lock script for connected inactive monitor
Previously wouldn't work (raised AttributeError) when a monitor was
connected but not active as the resolution was not reported, and thus
the regular expression failed to report resolution.

We now check for an Attribute error for monitors and just continue to
the next monitor. (Only should effect addition of lock image to monitor)